requirements.txt
pytest==7.2.1
selenium==4.8.2
pre-commit==3.1.0
allure-pytest==2.12.0
